Enable use of larger window for k+q than k in LBTE with symmetry
Previously, if a k point needed for interpolation of el_f (k+q) is not included
in el_i (k), vector_field_unfold_and_interpolate_map raised an error.
Now, we assume that if a state is not included in el_i, its delta f (occupation
change) is 0. So, we can neglect such points.